refactor(table): rename *-helper.ts composables to use*.ts convention
