+++ components/table/src/Table/useTable.ts
@@ -6,10 +6,10 @@ import { useLocale } from '@flash-global66/g-hooks'
 import { useNamespace } from '@flash-global66/g-utils'
 import { createStore } from '../shared/store/helper'
 import TableLayout from '../shared/composables/table-layout'
-import useUtils from './utils-helper'
-import { convertToRows } from '../components/TableHeader/utils-helper'
-import useStyle from './style-helper'
-import useKeyRender from './key-render-helper'
+import useUtils from './useUtils'
+import { convertToRows } from '../components/TableHeader/useUtils'
+import useStyle from './useStyle'
+import useKeyRender from './useKeyRender'
 import { TABLE_INJECTION_KEY } from '../shared/constants/token.constant'
 import { useScrollbar } from '../shared/composables/useScrollbar'
 import { clearExpansionCacheForTable } from '../components/TableColumn/cell-renderers/cell-expansion-utils'
