+++ components/table/src/components/TableBody/index.ts
@@ -13,7 +13,7 @@ import { useNamespace } from '@flash-global66/g-utils'
 import useLayoutObserver from '../../shared/composables/useLayoutObserver'
 import { removePopper } from '../../shared/utils/table.util'
 import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
-import useRender from './render-helper'
+import useRender from './useRender'
 import defaultProps from './defaults'
 
 import type { VNode } from 'vue'
