+++ components/table/src/components/TableBody/useRender.ts
@@ -6,8 +6,8 @@ import { useNamespace } from '@flash-global66/g-utils'
 import { isBoolean, isPropAbsent } from '@flash-global66/g-utils'
 import { getRowIdentity } from '../../shared/utils/table.util'
 import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
-import useEvents from './events-helper'
-import useStyles from './styles-helper'
+import useEvents from './useEvents'
+import useStyles from './useStyles'
 import TdWrapper from './td-wrapper.vue'
 import type { TableBodyProps } from './defaults'
 import type { RenderRowData, TableProps, TreeNode } from '../../Table/defaults'
