+++ components/table/src/components/TableColumn/index.ts
@@ -15,8 +15,8 @@ import GCheckbox from '@flash-global66/g-checkbox'
 import { isArray, isString, isUndefined } from '@flash-global66/g-utils'
 import { cellStarts } from '../../shared/constants/renderer.constant'
 import { compose, mergeOptions } from '../../shared/utils/table.util'
-import useWatcher from './watcher-helper'
-import useRender from './render-helper'
+import useWatcher from './useWatcher'
+import useRender from './useRender'
 import defaultProps from './defaults'
 import type { TableColumn, TableColumnCtx } from './defaults'
