+++ components/table/src/components/TableHeader/index.ts
@@ -16,9 +16,9 @@ import { useNamespace } from '@flash-global66/g-utils'
 import FilterPanel from '../FilterPanel/index.vue'
 import useLayoutObserver from '../../shared/composables/useLayoutObserver'
 import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
-import useEvent from './event-helper'
-import useStyle from './style.helper'
-import useUtils from './utils-helper'
+import useEvent from './useEvent'
+import useStyle from './useStyle'
+import useUtils from './useUtils'
 import type { ComponentInternalInstance, PropType, Ref } from 'vue'
 import type { DefaultRow, Sort } from '../../Table/defaults'
 import type { Store } from '../../shared/store'
